requirements.txt
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
urllib3>=2.0
